comsolidate panel and base methods
