+++ bayespecon/models/base.py
@@ -27,6 +27,7 @@ from .._logdet import (
     resolve_logdet_bounds,
 )
 from ._base._shared import (
+    SharedSpatialMethods,
     _parse_W,
     _pointwise_gaussian_loglik,
     _write_log_likelihood_to_idata,
@@ -34,7 +35,7 @@ from ._base._shared import (
 )
 
 
-class SpatialModel(ABC):
+class SpatialModel(SharedSpatialMethods, ABC):
     """Base class for Bayesian spatial regression models. Models follow the notation
     of :cite:p:`anselin1988SpatialEconometrics` and :cite:p:`lesage2009IntroductionSpatial`.
     The API supports both formula and matrix input modes.
@@ -447,38 +448,6 @@ class SpatialModel(ABC):
             weights=eigs * V_col_sums * c,
         )
 
-    @property
-    def _nonintercept_indices(self) -> list[int]:
-        """Return indices of non-constant (non-intercept) columns in X.
-
-        This is used to exclude the intercept from impact measures, since
-        the intercept has no meaningful spatial effect interpretation.
-
-        Returns
-        -------
-        list[int]
-            Column indices of X that are not constant/intercept columns.
-        """
-        indices: list[int] = []
-        for j, name in enumerate(self._feature_names):
-            column = self._X[:, j]
-            is_named_intercept = name.lower() == "intercept"
-            is_constant = np.allclose(column, column[0])
-            if not (is_named_intercept or is_constant):
-                indices.append(j)
-        return indices
-
-    @property
-    def _nonintercept_feature_names(self) -> list[str]:
-        """Return feature names for non-intercept columns.
-
-        Returns
-        -------
-        list[str]
-            Feature names excluding intercept/constant columns.
-        """
-        return [self._feature_names[i] for i in self._nonintercept_indices]
-
     # ------------------------------------------------------------------
     # Input parsing helpers
     # ------------------------------------------------------------------
@@ -592,75 +561,9 @@ class SpatialModel(ABC):
     # Abstract interface
     # ------------------------------------------------------------------
 
-    def _add_nu_prior(self, model: pm.Model) -> pm.Model:
-        """Add the degrees-of-freedom prior for robust (Student-t) models.
-
-        Called inside ``_build_pymc_model`` when ``self.robust`` is True.
-        Uses an :math:`\\mathrm{Exp}(\\lambda_\\nu)` prior on ``nu`` with rate ``nu_lam`` (default
-        1/30, giving mean ≈ 30, favouring near-Normal tails). A lower
-        bound of 2 is enforced so that the variance exists.
-
-        Parameters
-        ----------
-        model : pymc.Model
-            The model context in which to add the ``nu`` prior.
-
-        Returns
-        -------
-        pymc.Model
-            The same model context (``nu`` is added as a side effect).
-        """
-        nu_lam = self.priors.get("nu_lam", 1.0 / 30.0)
-        pm.Truncated("nu", pm.Exponential.dist(lam=nu_lam), lower=2.0)
-        return model
-
     @abstractmethod
     def _build_pymc_model(self) -> pm.Model:
         """Construct and return a pm.Model. Subclasses implement this."""
-
-    def _beta_names(self) -> list[str]:
-        """Return coefficient labels used for posterior summaries.
-
-        Returns
-        -------
-        list[str]
-            Coefficient labels aligned with the ``beta`` parameter.
-        """
-        return list(self._feature_names)
-
-    def _model_coords(self) -> dict[str, list[str]]:
-        """Return PyMC coordinate labels for named dimensions.
-
-        Returns
-        -------
-        dict[str, list[str]]
-            Coordinates passed to :class:`pymc.Model`.
-        """
-        return {"coefficient": self._beta_names()}
-
-    @staticmethod
-    def _rename_summary_index(summary_df: pd.DataFrame) -> pd.DataFrame:
-        """Strip the ``beta[...]`` wrapper from coefficient row labels.
-
-        Parameters
-        ----------
-        summary_df : pandas.DataFrame
-            ArviZ summary output.
-
-        Returns
-        -------
-        pandas.DataFrame
-            Summary with human-readable coefficient row labels.
-        """
-        renamed = []
-        for label in summary_df.index.astype(str):
-            if label.startswith("beta[") and label.endswith("]"):
-                renamed.append(label[5:-1])
-            else:
-                renamed.append(label)
-        out = summary_df.copy()
-        out.index = renamed
-        return out
 
     # ------------------------------------------------------------------
     # Public API
@@ -1130,18 +1033,6 @@ class SpatialModel(ABC):
             )
 
         idata["log_likelihood"] = xr.Dataset(new_vars)
-
-    @property
-    def inference_data(self) -> Optional[az.InferenceData]:
-        """Return the ArviZ InferenceData from the most recent fit.
-
-        Returns
-        -------
-        arviz.InferenceData or None
-            The inference data object, or ``None`` if the model has not
-            been fit yet.
-        """
-        return self._idata
 
     @property
     def pymc_model(self) -> Optional[pm.Model]:
@@ -1738,10 +1629,6 @@ class SpatialModel(ABC):
     # Internals
     # ------------------------------------------------------------------
 
-    def _require_fit(self):
-        if self._idata is None:
-            raise RuntimeError("Model has not been fit yet. Call .fit() first.")
-
     def _require_W(self):
         """Raise if no spatial weights matrix was supplied."""
         if self._W_sparse is None:
@@ -1749,9 +1636,6 @@ class SpatialModel(ABC):
                 "This method requires a spatial weights matrix W. "
                 "Pass W when constructing the model."
             )
-
-    def _posterior_mean(self, var: str) -> np.ndarray:
-        return self._idata.posterior[var].mean(("chain", "draw")).to_numpy()
 
     def fitted_values(self) -> np.ndarray:
         """Return fitted values at posterior mean parameters.
