+++ bayespecon/models/panel_base.py
@@ -26,7 +26,7 @@ from .._logdet import (
     make_logdet_numpy_fn,
     make_logdet_numpy_vec_fn,
 )
-from ._base._shared import _is_row_standardized_csr
+from ._base._shared import SharedSpatialMethods, _is_row_standardized_csr
 from .base import (
     _pointwise_gaussian_loglik,
     _write_log_likelihood_to_idata,
@@ -254,7 +254,7 @@ def _parse_panel_W(
     return W_csr, row_std
 
 
-class SpatialPanelModel(ABC):
+class SpatialPanelModel(SharedSpatialMethods, ABC):
     """Base class for static spatial panel models with FE transforms.
 
     Holds the within-transformation, panel-aware sorting, and weights
@@ -823,38 +823,6 @@ class SpatialPanelModel(ABC):
         return _chunked_eig_means(rho_draws, eigs, weights=eigs * V_col_sums * c)
 
     @property
-    def _nonintercept_indices(self) -> list[int]:
-        """Return indices of non-constant (non-intercept) columns in X.
-
-        This is used to exclude the intercept from impact measures, since
-        the intercept has no meaningful spatial effect interpretation.
-
-        Returns
-        -------
-        list[int]
-            Column indices of X that are not constant/intercept columns.
-        """
-        indices: list[int] = []
-        for j, name in enumerate(self._feature_names):
-            column = self._X[:, j]
-            is_named_intercept = name.lower() == "intercept"
-            is_constant = np.allclose(column, column[0])
-            if not (is_named_intercept or is_constant):
-                indices.append(j)
-        return indices
-
-    @property
-    def _nonintercept_feature_names(self) -> list[str]:
-        """Return feature names for non-intercept columns.
-
-        Returns
-        -------
-        list[str]
-            Feature names excluding intercept/constant columns.
-        """
-        return [self._feature_names[i] for i in self._nonintercept_indices]
-
-    @property
     def _intercept_dropped(self) -> bool:
         """Whether the intercept column was dropped from the posterior beta.
 
@@ -946,28 +914,6 @@ class SpatialPanelModel(ABC):
                 indices.append(j)
         return indices
 
-    def _add_nu_prior(self, model: pm.Model) -> pm.Model:
-        """Add the degrees-of-freedom prior for robust (Student-t) models.
-
-        Called inside ``_build_pymc_model`` when ``self.robust`` is True.
-        Uses an :math:`\\mathrm{Exp}(\\lambda_\\nu)` prior on ``nu`` with rate ``nu_lam`` (default
-        1/30, giving mean ≈ 30, favouring near-Normal tails). A lower
-        bound of 2 is enforced so that the variance exists.
-
-        Parameters
-        ----------
-        model : pymc.Model
-            The model context in which to add the ``nu`` prior.
-
-        Returns
-        -------
-        pymc.Model
-            The same model context (``nu`` is added as a side effect).
-        """
-        nu_lam = self.priors.get("nu_lam", 1.0 / 30.0)
-        pm.Truncated("nu", pm.Exponential.dist(lam=nu_lam), lower=2.0)
-        return model
-
     @abstractmethod
     def _build_pymc_model(self) -> pm.Model:
         """Construct and return a pm.Model."""
@@ -989,50 +935,6 @@ class SpatialPanelModel(ABC):
     @abstractmethod
     def _fitted_mean_from_posterior(self) -> np.ndarray:
         """Posterior-mean fitted values on transformed scale."""
-
-    def _beta_names(self) -> list[str]:
-        """Return coefficient labels used for posterior summaries.
-
-        Returns
-        -------
-        list[str]
-            Coefficient labels aligned with the ``beta`` parameter.
-        """
-        return list(self._feature_names)
-
-    def _model_coords(self) -> dict[str, list[str]]:
-        """Return PyMC coordinate labels for named dimensions.
-
-        Returns
-        -------
-        dict[str, list[str]]
-            Coordinates passed to :class:`pymc.Model`.
-        """
-        return {"coefficient": self._beta_names()}
-
-    @staticmethod
-    def _rename_summary_index(summary_df: pd.DataFrame) -> pd.DataFrame:
-        """Strip the ``beta[...]`` wrapper from coefficient row labels.
-
-        Parameters
-        ----------
-        summary_df : pandas.DataFrame
-            ArviZ summary output.
-
-        Returns
-        -------
-        pandas.DataFrame
-            Summary with human-readable coefficient row labels.
-        """
-        renamed = []
-        for label in summary_df.index.astype(str):
-            if label.startswith("beta[") and label.endswith("]"):
-                renamed.append(label[5:-1])
-            else:
-                renamed.append(label)
-        out = summary_df.copy()
-        out.index = renamed
-        return out
 
     def fit(
         self,
@@ -1469,18 +1371,6 @@ class SpatialPanelModel(ABC):
         idata["log_likelihood"] = xr.Dataset(new_vars)
 
     @property
-    def inference_data(self) -> Optional[az.InferenceData]:
-        """Return the ArviZ InferenceData from the most recent fit.
-
-        Returns
-        -------
-        arviz.InferenceData or None
-            The inference data object, or ``None`` if the model has not
-            been fit yet.
-        """
-        return self._idata
-
-    @property
     def pymc_model(self) -> Optional[pm.Model]:
         """Return the PyMC model object built for the most recent fit.
 
@@ -1491,13 +1381,6 @@ class SpatialPanelModel(ABC):
             has not been fit yet.
         """
         return self._pymc_model
-
-    def _require_fit(self):
-        if self._idata is None:
-            raise RuntimeError("Model has not been fit yet. Call .fit() first.")
-
-    def _posterior_mean(self, var: str) -> np.ndarray:
-        return self._idata.posterior[var].mean(("chain", "draw")).to_numpy()
 
     def summary(self, var_names: Optional[list] = None, **kwargs) -> pd.DataFrame:
         """Return posterior summary table.
